feat: initial extraction of input.rs
